New signing keys, to fix automatic update
There's a bug in the key generation script included with the version of
Sparkle that I'm using. It was fixed here:
https://github.com/andymatuschak/Sparkle/pull/15

When I was initally testing sparkle updates, things worked fine, but
apparently something must have changed with my OpenSSL version because
both versions I have won't generate keys that the old version of
MP3Gain Express will accept.

I've generated new keys using the fixed versions of the script, and
updated MP3Gain Express to 1.1.1 to use the new public key. The 1.1
download was updated to include the new key as well, so anyone who
downloads it now using an old link will properly be able to update to
1.1.1. (Which is identical)